Rename duke to chrisPBacon
