+++ README.md
@@ -1,4 +1,4 @@
-# seedu.duke.seedu.duke project template
+# seedu.chrisPBacon.seedu.chrisPBacon project template
 
 This is a project template for a greenfield Java project. It's named after the Java mascot _Duke_. Given below are instructions on how to use it.
 
@@ -13,7 +13,7 @@ Prerequisites: JDK 11, update Intellij to the most recent version.
    1. If there are any further prompts, accept the defaults.
 1. Configure the project to use **JDK 11** (not other versions) as explained in [here](https://www.jetbrains.com/help/idea/sdk.html#set-up-jdk).<br>
    In the same dialog, set the **Project language level** field to the `SDK default` option.
-3. After that, locate the `src/main/java/seedu.duke.seedu.duke.java` file, right-click it, and choose `Run seedu.duke.seedu.duke.main()` (if the code editor is showing compile errors, try restarting the IDE). If the setup is correct, you should see something like the below as the output:
+3. After that, locate the `src/main/java/seedu.chrisPBacon.seedu.chrisPBacon.java` file, right-click it, and choose `Run seedu.chrisPBacon.seedu.chrisPBacon.main()` (if the code editor is showing compile errors, try restarting the IDE). If the setup is correct, you should see something like the below as the output:
    ```
    Hello from
     ____        _
